#3369: moving files back to correct folder (this should work)
