+++ src/frontend/src/pages/PartPage/components/PartDisplay.tsx
