+++ src/frontend/src/pages/PartPage/components/PartOverview.tsx
