+++ src/frontend/src/pages/PartPage/components/PartSubmissionDetails.tsx
