+++ src/frontend/src/pages/PartPage/components/PdfDisplay.tsx
